requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
